refactor: split webhook.go into 3 files by concern (#57)
Split webhook.go (652 lines) into:
- webhook.go (292 lines) — core output lifecycle, batch goroutine
- webhook_config.go (232 lines) — config struct, validation, defaults, TLS
- webhook_http.go (173 lines) — HTTP POST, retry, NDJSON, backoff

All under 300-line guideline. No API or behaviour changes.
